requirements.txt
_ipyw_jlab_nb_ext_conf	0.1.0	0.1.0
absl-py	0.9.0	0.13.0
aiohttp	3.6.2	3.7.4.post0
alabaster	0.7.12	0.7.12
anaconda	2019.10	2021.05
anaconda-client	1.7.2	1.9.0
anaconda-navigator	1.9.7	2.1.0
anaconda-project	0.8.3	0.10.1
asn1crypto	1.0.1	1.4.0
astroid	2.3.1	2.7.1
astropy	3.2.1	4.3.1
async-timeout	3.0.1	3.0.1
atomicwrites	1.3.0	1.4.0
attrs	19.2.0	21.2.0
babel	2.7.0	2.9.1
backcall	0.1.0	0.2.0
backports	1.0	1.0
backports.functools_lru_cache	1.5	1.6.4
backports.os	0.1.1	0.1.1
backports.shutil_get_terminal_size	1.0.0	1.0.0
backports.tempfile	1.0	1.0
backports.weakref	1.0.post1	1.0.post1
beautifulsoup4	4.8.0	4.10.0
bitarray	1.0.1	2.3.0
bkcharts	0.2	0.2
blas	1.0	1.0
bleach	3.1.0	4.0.0
blosc	1.16.3	1.21.0
bokeh	1.3.4	2.4.1
boto	2.49.0	2.49.0
bottleneck	1.2.1	1.3.2
bzip2	1.0.8	1.0.8
ca-certificates	2019.8.28	2021.10.26
certifi	2019.9.11	2021.10.8
cffi	1.12.3	1.14.6
chardet	3.0.4	4.0.0
click	7.0	8.0.3
cloudpickle	1.2.2	2.0.0
clyent	1.2.2	1.2.2
colorama	0.4.1	0.4.4
comtypes	1.1.7	1.1.10
conda	4.8.2	4.10.3
conda-build	3.18.9	3.21.5
conda-env	2.6.0	2.6.0
conda-package-handling	1.6.0	1.7.3
conda-verify	3.4.2	3.4.2
console_shortcut	0.1.1	0.1.1
contextlib2	0.6.0	0.6.0.post1
cryptography	2.7	35.0.0
cudatoolkit	10.1.243	11.3.1
curl	7.65.3	7.78.0
cycler	0.10.0	0.10.0
cython	0.29.13	0.29.24
cytoolz	0.10.0	0.11.0
dask	2.5.2	2021.10.0
dask-core	2.5.2	2021.10.0
decorator	4.4.0	5.1.0
deepdiff	4.2.0	
defusedxml	0.6.0	0.7.1
distributed	2.5.2	2021.10.0
docutils	0.15.2	0.17.1
dominate	2.6.0	
efficientnet-pytorch	0.7.0	
entrypoints	0.3	0.3
enum34	1.1.9	1.1.10
et_xmlfile	1.0.1	1.1.0
fastcache	1.1.0	1.1.0
filelock	3.0.12	3.3.1
flask	1.1.1	2.0.2
freetype	2.9.1	2.10.4
fsspec	0.5.2	2021.8.1
future	0.17.1	0.18.2
get_terminal_size	1.0.0	1.0.0
gevent	1.4.0	21.8.0
glob2	0.7	0.7
greenlet	0.4.15	1.1.1
h5py	2.9.0	3.2.1
hdf5	1.10.4	1.12.0
heapdict	1.0.1	1.0.1
html5lib	1.0.1	1.1
icc_rt	2019.0.0	2019.0.0
icu	58.2	68.1
idna	2.8	3.2
imageio	2.6.0	2.9.0
imagesize	1.1.0	1.2.0
importlib_metadata	0.23	4.8.1
intel-openmp	2019.4	2021.4.0
ipykernel	5.1.2	6.4.1
ipython	7.8.0	7.29.0
ipython_genutils	0.2.0	0.2.0
ipywidgets	7.5.1	7.6.5
isort	4.3.21	5.9.3
itsdangerous	1.1.0	2.0.1
jdcal	1.4.1	1.4.1
jedi	0.15.1	0.18.0
jinja2	2.10.3	3.0.2
joblib	0.13.2	1.1.0
jpeg	9b	9d
json5	0.8.5	0.9.6
jsonpatch	1.32	1.32
jsonpointer	2.1	2.1
jsonschema	3.0.2	3.2.0
jupyter	1.0.0	1.0.0
jupyter_client	5.3.3	7.0.1
jupyter_console	6.0.0	6.4.0
jupyter_core	4.5.0	4.8.1
jupyterlab	1.1.4	3.2.1
jupyterlab_server	1.0.6	2.8.2
keyring	18.0.0	23.1.0
kiwisolver	1.1.0	1.3.1
krb5	1.16.1	1.19.2
lazy-object-proxy	1.4.2	1.6.0
libarchive	3.3.3	3.4.2
libcurl	7.65.3	7.78.0
libiconv	1.15	1.15
liblief	0.9.0	0.10.1
libpng	1.6.37	1.6.37
libsodium	1.0.16	1.0.18
libssh2	1.8.2	1.9.0
libtiff	4.0.10	4.2.0
libxml2	2.9.9	2.9.12
libxslt	1.1.33	1.1.34
llvmlite	0.29.0	0.37.0
locket	0.2.0	0.2.1
lxml	4.4.1	4.6.3
lz4-c	1.8.1.2	1.9.3
lzo	2.10	2.10
m2w64-gcc-libgfortran	5.3.0	5.3.0
m2w64-gcc-libs	5.3.0	5.3.0
m2w64-gcc-libs-core	5.3.0	5.3.0
m2w64-gmp	6.1.0	6.1.0
m2w64-libwinpthread-git	5.0.0.4634.697f757	5.0.0.4634.697f757
markupsafe	1.1.1	2.0.1
matplotlib	3.1.1	3.4.3
mccabe	0.6.1	0.6.1
menuinst	1.4.16	1.4.18
mistune	0.8.4	0.8.4
mkl	2019.4	2021.4.0
mkl-service	2.3.0	2.4.0
mkl_fft	1.0.14	1.3.1
mkl_random	1.1.0	1.2.2
mock	3.0.5	4.0.3
more-itertools	7.2.0	8.10.0
mpmath	1.1.0	1.2.1
mpyq	0.2.5	
msgpack-python	0.6.1	1.0.2
msys2-conda-epoch	20160418	20160418
multidict	4.7.5	5.1.0
multipledispatch	0.6.0	0.6.0
natsort	7.1.1	7.1.1
navigator-updater	0.2.1	0.2.1
nbconvert	5.6.0	6.1.0
nbformat	4.4.0	5.1.3
networkx	2.3	2.6.3
nibabel	3.2.1	
ninja	1.7.2	1.10.2
nltk	3.4.5	3.6.5
nose	1.3.7	1.3.7
notebook	6.0.1	6.4.5
numba	0.45.1	0.54.1
numexpr	2.7.0	2.7.3
numpy	1.16.5	1.21.2
numpy-base	1.16.5	1.21.2
numpydoc	0.9.1	1.1.0
olefile	0.46	0.46
opencv-python	4.2.0.32	
openpyxl	3.0.0	3.0.9
openssl	1.1.1d	1.1.1l
ordered-set	3.1.1	
packaging	19.2	21.0
pandas	0.25.1	1.3.4
pandoc	2.2.3.2	2.12
pandocfilters	1.4.2	1.4.3
parso	0.5.1	0.8.2
partd	1.0.0	1.2.0
path.py	12.0.1	12.5.0
pathlib2	2.3.5	2.3.6
patsy	0.5.1	0.5.2
pep8	1.7.1	1.7.1
pgmagick	0.7.5	
pickleshare	0.7.5	0.7.5
pillow	6.2.0	8.4.0
pip	20.1	21.2.4
pkginfo	1.5.0.1	1.7.1
pluggy	0.13.0	1.0.0
ply	3.11	3.11
portpicker	1.3.1	1.3.1
powershell_shortcut	0.0.1	0.0.1
prometheus_client	0.7.1	0.11.0
prompt_toolkit	2.0.10	3.0.20
protobuf	3.11.3	3.17.2
psutil	5.6.3	5.8.0
py	1.8.0	1.10.0
py-lief	0.9.0	0.10.1
pybrisque	1.0	
pycodestyle	2.5.0	2.7.0
pycosat	0.6.3	0.6.3
pycparser	2.19	2.20
pycrypto	2.6.1	2.6.1
pycurl	7.43.0.3	7.44.1
pydicom	2.1.2	
pyflakes	2.1.1	2.3.1
pygame	1.9.6	
pyglet	1.5.0	
pygments	2.4.2	2.10.0
pyhook	1.5.1	
pylint	2.4.2	2.9.6
pymouse	1.0	
pyodbc	4.0.27	4.0.31
pyopenssl	19.0.0	21.0.0
pyparsing	2.4.2	3.0.4
pyqt	5.9.2	5.9.2
pyreadline	2.1	2.1
pyrsistent	0.15.4	0.18.0
pysc2	3.0.0	
pysocks	1.7.1	1.7.1
pytables	3.5.2	3.6.1
pytest	5.2.1	6.2.4
pytest-arraydiff	0.3	0.3
pytest-astropy	0.5.0	0.8.0
pytest-doctestplus	0.4.0	0.10.1
pytest-openfiles	0.4.0	0.5.0
pytest-remotedata	0.3.2	0.3.2
python	3.7.4	3.10.0
python-dateutil	2.8.0	2.8.2
python-libarchive-c	2.8	2.9
pytorch	1.4.0	1.6.0
pytorch-ignite	0.3.0	
pytz	2019.3	2021.3
pyuserinput	0.1.11	
pywavelets	1.0.3	1.1.1
pywin32	223	228
pywinpty	0.5.5	0.5.7
pyyaml	5.1.2	6.0
pyzmq	18.1.0	22.2.1
qt	5.9.7	5.9.7
qtawesome	0.6.0	1.0.2
qtconsole	4.5.5	5.1.1
qtpy	1.9.0	1.10.0
requests	2.22.0	2.26.0
rope	0.14.0	0.19.0
ruamel_yaml	0.15.46	0.15.100
s2clientprotocol	4.11.4.78285.0	
s2protocol	4.11.4.78285.0	
sc2	0.11.2	
scikit-image	0.15.0	0.18.3
scikit-learn	0.21.3	1.0.1
scipy	1.2.0	1.7.1
seaborn	0.9.0	0.11.2
send2trash	1.5.0	1.8.0
setuptools	45.2.0	58.0.4
simplegeneric	0.8.1	0.8.1
simpleitk	2.0.2	
singledispatch	3.4.0.3	3.7.0
sip	4.19.8	4.19.25
six	1.12.0	1.16.0
sk-video	1.1.10	
snappy	1.1.7	1.1.8
snowballstemmer	2.0.0	2.1.0
sortedcollections	1.1.2	2.1.0
sortedcontainers	2.1.0	2.4.0
soupsieve	1.9.3	2.2.1
sphinx	2.2.0	4.2.0
sphinxcontrib	1.0	1.0
sphinxcontrib-applehelp	1.0.1	1.0.2
sphinxcontrib-devhelp	1.0.1	1.0.2
sphinxcontrib-htmlhelp	1.0.2	2.0.0
sphinxcontrib-jsmath	1.0.1	1.0.1
sphinxcontrib-qthelp	1.0.2	1.0.3
sphinxcontrib-serializinghtml	1.1.3	1.1.5
sphinxcontrib-websupport	1.1.2	1.2.4
spyder	3.3.6	5.1.5
spyder-kernels	0.5.2	2.1.3
sqlalchemy	1.3.9	1.4.22
sqlite	3.30.0	3.36.0
statsmodels	0.10.1	0.13.0
sympy	1.4	1.9
tbb	2019.4	2021.4.0
tblib	1.4.0	1.7.0
tensorboardx	2.0	2.2
terminado	0.8.2	0.9.4
testpath	0.4.2	0.5.0
thop	0.0.31-2005241907	
tk	8.6.8	8.6.11
toolz	0.10.0	0.11.1
torchfile	0.1.0	
torchsummary	1.5.1	
torchvision	0.5.0	
tornado	6.0.3	6.1
tqdm	4.36.1	4.62.3
traitlets	4.3.3	5.1.0
unicodecsv	0.14.1	0.14.1
urllib3	1.24.2	1.26.7
vc	14.1	14.2
visdom	0.1.8.9	
vs2015_runtime	14.16.27012	14.27.29016
wcwidth	0.1.7	0.2.5
webencodings	0.5.1	0.5.1
websocket-client	0.57.0	0.58.0
werkzeug	0.16.0	2.0.2
wheel	0.33.6	0.37.0
whichcraft	0.6.1	0.6.1
widgetsnbextension	3.5.1	3.5.1
win_inet_pton	1.1.0	1.1.0
win_unicode_console	0.5	0.5
wincertstore	0.2	0.2
winpty	0.4.3	0.4.3
wrapt	1.11.2	1.12.1
xlrd	1.2.0	2.0.1
xlsxwriter	1.2.1	3.0.1
xlwings	0.15.10	0.24.9
xlwt	1.3.0	1.3.0
xz	5.2.4	5.2.5
yaml	0.1.7	0.2.5
yarl	1.4.2	1.6.3
zeromq	4.3.1	4.3.4
zict	1.0.0	2.0.0
zipp	0.6.0	3.6.0
zlib	1.2.11	1.2.11